Add Resilience4j documentation
The documentation used to describe how to use Hystrix with Vert.x.
Since Hystrix has been obsolete for a long time and the documentation
was removed in a recent commit, this commit adds a modern replacement:
Resilience4j. The documentation mainly links to the Resilience4j
Vert.x how-to that was published recently.